+++ tests/Type/GeneralTypeTest.php
@@ -22,6 +22,7 @@ class GeneralTypeTest extends TypeInferenceTestCase
         yield from $this->gatherAssertTypes(__DIR__.'/data/conditionable.php');
         yield from $this->gatherAssertTypes(__DIR__.'/data/translate.php');
         yield from $this->gatherAssertTypes(__DIR__.'/data/model-factories.php');
+        yield from $this->gatherAssertTypes(__DIR__.'/data/environment-helper.php');
     }
 
     /**